hf_space_emotion_api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pillow==10.4.0
numpy==1.26.4
onnxruntime==1.18.1
python-multipart==0.0.9
